requirements.txt
beautifulsoup4==4.9.3
databases==0.5.3
Django==4.0.3
django-allauth==0.51.0
django-bootstrap5==21.3
django-colorfield==0.7.1
django-extensions==3.1.5
django-pygments-renderer==0.0.1
defusedxml==0.7.1
fake-useragent==0.1.11
google-ads==15.0.0
google-analytics-data==0.11.1
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleads==31.0.0
googleapis-common-protos==1.56.0
Jinja2==3.0.2
lxml==4.6.3
oauth2client==4.1.3
oauthlib==3.1.1
paramiko==2.9.2
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
selenium==3.141.0
selenium-stealth==1.0.6
SQLAlchemy==1.4.27
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.13.4
webdriver-manager==3.4.2
Werkzeug==2.0.2
